Move openrouter and langchain helpers to router
